add us0002-n

+++ socialMeli/src/main/java/com/javabootcamp/socialmeli/repository/FollowRepository.java
@@ -10,5 +10,5 @@ public interface FollowRepository {
     void delete(Integer followerId, Integer followedId);
     List<User> findFollowersById(Integer id);
     List<User> findFollowedsById(Integer id);
-    Long countFollowersById(Integer id);
+    int countFollowersById(Integer id);
 }
